Set `s3_is_enabled` only if ACP input passes all checks

+++ acp/main_module.php
@@ -36,38 +36,50 @@ class main_module
                 $errors[] = $user->lang('ACP_S3_AWS_SECRET_ACCESS_KEY_INVALID', $request->variable('s3_aws_secret_access_key', ''));
             }
 
-            try {
-                // Instantiate an AWS S3 client.
-                $s3_client = new \Aws\S3\S3Client([
-                    'credentials' => [
-                        'key'    => $request->variable('s3_aws_access_key_id', ''),
-                        'secret' => $request->variable('s3_aws_secret_access_key', ''),
-                    ],
-                    'debug'       => false,
-                    'http'        => [
-                        'verify' => false,
-                    ],
-                    'region'      => $request->variable('s3_region', ''),
-                    'version'     => 'latest',
-                ]);
-
-                // Upload a test file to ensure credentials are valid and everything is working properly.
-                $s3_client->upload($request->variable('s3_bucket', ''), 'test.txt', 'test body');
-
-                // Delete the test file.
-                $s3_client->deleteObject([
-                    'Bucket' => $request->variable('s3_bucket', ''),
-                    'Key'    => 'test.txt',
-                ]);
-            } catch (\Aws\S3\Exception\S3Exception $e) {
-                $errors[] = $e->getMessage();
+            if (empty($request->variable('s3_region', ''))) {
+                $errors[] = $user->lang('ACP_S3_REGION_INVALID');
             }
 
+            if (empty($request->variable('s3_bucket', ''))) {
+                $errors[] = $user->lang('ACP_S3_BUCKET_INVALID');
+            }
+
+            // If we have no errors so far, let's ensure our AWS credentials are actually working.
+            if (!count($errors)) {
+                try {
+                    // Instantiate an AWS S3 client.
+                    $s3_client = new \Aws\S3\S3Client([
+                        'credentials' => [
+                            'key'    => $request->variable('s3_aws_access_key_id', ''),
+                            'secret' => $request->variable('s3_aws_secret_access_key', ''),
+                        ],
+                        'http'        => [
+                            'verify' => false,
+                        ],
+                        'region'      => $request->variable('s3_region', ''),
+                        'version'     => 'latest',
+                    ]);
+
+                    // Upload a test file to ensure credentials are valid and everything is working properly.
+                    $s3_client->upload($request->variable('s3_bucket', ''), 'test.txt', 'test body');
+
+                    // Delete the test file.
+                    $s3_client->deleteObject([
+                        'Bucket' => $request->variable('s3_bucket', ''),
+                        'Key'    => 'test.txt',
+                    ]);
+                } catch (\Aws\S3\Exception\S3Exception $e) {
+                    $errors[] = $e->getMessage();
+                }
+            }
+
+            // If we still don't have any errors, it is time to set the database config values.
             if (!count($errors)) {
                 $config->set('s3_aws_access_key_id', $request->variable('s3_aws_access_key_id', ''));
                 $config->set('s3_aws_secret_access_key', $request->variable('s3_aws_secret_access_key', ''));
                 $config->set('s3_region', $request->variable('s3_region', ''));
                 $config->set('s3_bucket', $request->variable('s3_bucket', ''));
+                $config->set('s3_is_enabled', 1);
 
                 trigger_error($user->lang('ACP_S3_SETTING_SAVED') . adm_back_link($this->u_action));
             }
@@ -80,6 +92,7 @@ class main_module
             'S3_AWS_SECRET_ACCESS_KEY' => $config['s3_aws_secret_access_key'],
             'S3_REGION'                => $config['s3_region'],
             'S3_BUCKET'                => $config['s3_bucket'],
+            'S3_IS_ENABLED'            => ($config['s3_is_enabled']) ? 'Enabled' : 'Disabled',
         ]);
     }
 }
